Moved all code under src
