+++ src/volttron/client/vip/green.py
@@ -54,7 +54,7 @@ from gevent.lock import RLock as _RLock
 from zmq.green import NOBLOCK, POLLOUT
 from zmq import green as _green
 
-from src.volttron.utils.socket import _Socket
+from volttron.utils.socket import _Socket
 # from volttron.client.vip.router import BaseRouter as _BaseRouter
