+++ src/volttron/server/router/green_router.py
@@ -3,7 +3,7 @@ from typing import Optional
 import zmq
 from zmq import green as _green
 
-from src.volttron.services.peer import ServicePeerNotifier
+from volttron.services.peer import ServicePeerNotifier
 from .router import Router
