+++ tests/utils/test_sqlite3_datetime.py
@@ -37,7 +37,7 @@
 # }}}
 
 from dateutil.parser import parse
-from src.volttron.utils import fix_sqlite3_datetime
+from volttron.utils import fix_sqlite3_datetime
 import sqlite3 as sql
